% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{check_cvd_kcd}
\alias{check_cvd_kcd}
\title{Check covered kcd code}
\usage{
check_cvd_kcd(cvd_kcd, kcd)
}
\arguments{
\item{cvd_kcd}{covered kcd code vector}

\item{kcd}{kcd code vector}
}
\value{
the return type is a binary matrix
}
\description{
This function checks the covered kcd code from the kcd code vector
}
\examples{
head(check_cvd_kcd(c("A00", "A01", "A02"), kcd_book$kcd), 20)
}
