% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{check_cvd_level}
\alias{check_cvd_level}
\title{Check surgery level}
\usage{
check_cvd_level(cvd_level, level)
}
\arguments{
\item{level}{covered surgery level}

\item{kcd}{surgery level vector}
}
\value{
the return type is a binary matrix
}
\description{
This function checks the covered surgery level from the surgery level vector
}
\examples{
head(check_sur_lvl(c(1, 2, 3, 4, 5), c(1, 1, 3, 2, 4, 5, 2, 3, 2, 5)), 20)
}
