% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{count_stay}
\alias{count_stay}
\title{count stay}
\usage{
count_stay(df, id_var, from_var, to_var)
}
\arguments{
\item{df}{data.frame, data.table}

\item{id_var}{id variables}

\item{from_var}{start date}

\item{to_var}{end date}
}
\description{
This function count unique length of stay between \code{from_var} to \code{to_var}.
}
