% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{di_cell}
\alias{di_cell}
\title{Distance between two cells}
\usage{
di_cell(x, y)
}
\arguments{
\item{x}{a cell (a1 style)}

\item{y}{a cell (a1 style)}
}
\description{
This function calculates distance between two cells
}
\examples{
di_cell(x = "A1", y = "C3")
}
