% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{diff_period}
\alias{diff_period}
\title{Differences between start and end}
\usage{
diff_period(start, end, group = 1)
}
\arguments{
\item{start}{A date Date}

\item{end}{A date Date}

\item{group}{A numeric period}
}
\value{
Period matrix
}
\description{
Differences between start and end
}
