% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{get_rp_table}
\alias{get_rp_table}
\title{Get risk premium table in the excel file}
\usage{
get_rp_table(xlsxFile, sheet, risk_range, rate_range, skipEmptyRows = TRUE)
}
\arguments{
\item{xlsxFile}{file name character}

\item{sheet}{character}

\item{risk_rng}{vector}

\item{rate_rng}{vector}
}
\description{
This function convert the risk premium table in the excel file to a new format
}
\examples{
\dontrun{
get_rp_table(xlsxFile, sheet, c("D2", "S2"), c("D3", "S104"))
}
}
