% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{is_english}
\alias{is_english}
\alias{is_korean}
\alias{is_japanese}
\title{Is x English?}
\usage{
is_english(x)

is_korean(x)

is_japanese(x)
}
\arguments{
\item{x}{A string vector}
}
\description{
Is x English?
}
