% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join}
\alias{join}
\title{Merge data frames}
\usage{
join(..., by, all = FALSE, all.x = all, all.y = all, sort = TRUE)
}
\arguments{
\item{...}{data frames, or objects to be coerced to one.}

\item{by}{specifications of the columns used for merging.}

\item{all}{logical; all = L is shorthand for all.x = L and all.y = L, where L is either TRUE or FALSE.}

\item{all.y}{logical; analogous to all.x.}

\item{sort}{logical. Should the result be sorted on the by columns?}
}
\description{
This function merges several data frames at once
}
