% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{mv_cell}
\alias{mv_cell}
\title{Add rows and columns from a cell}
\usage{
mv_cell(x, r, c)
}
\arguments{
\item{x}{a cell (a1 style)}

\item{r}{row}

\item{c}{col}
}
\description{
This function converts cell range to numeric rows and columns values
}
\examples{
nu_cell(x = "A1", r = 1, c = 3)
}
