% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{nu_cell}
\alias{nu_cell}
\title{Numeric value of columns, rows}
\usage{
nu_cell(x, y)
}
\arguments{
\item{x}{a cell (a1 style)}

\item{y}{a cell (a1 style)}
}
\description{
This function converts cell range to numeric rows and columns values
}
\examples{
nu_cell(x = "A1", y = "C3")
}
