% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{overlap_matrix}
\alias{overlap_matrix}
\title{Overlap matrix}
\usage{
overlap_matrix(clm, clm_1st, col)
}
\arguments{
\item{clm}{a matrix. Claim matrix}

\item{clm_1st}{a matrix. Claim matrix only the first claim is 1 and the rest are 0 for each column and each row id}

\item{col}{a binary vector. it is 1 if it is one-time payment, it is 0 if it is continuous payment}
}
\description{
This function overlaps two matrices with same dimension by a binary vector.\cr
There are two claim methods, one for one-time payment and the other for continuous payments.\cr
Therefore, we have to combine the two matrices separately according to whether it is a one-time payment or continuous payments.
}
