% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proc_time}
\alias{proc_time}
\title{Simple running time of {vuw} package}
\usage{
proc_time(expr)
}
\arguments{
\item{expr}{Valid \R expression to be timed}
}
\description{
Simple running time of {vuw} package
}
\examples{
proc_time(df <- iris)
}
