% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{pull_cvd_kcd}
\alias{pull_cvd_kcd}
\title{Pull covered kcd code}
\usage{
pull_cvd_kcd(cvd_kcd, kcd)
}
\arguments{
\item{cvd_kcd}{covered kcd code. regular expression}

\item{kcd}{kcd code vector}
}
\value{
the return type is a character vector
}
\description{
This function pulled the covered kcd code from the kcd code vector
}
\examples{
pull_cvd_kcd("A05|B03", kcd_book$code)
}
