% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{reprow}
\alias{reprow}
\alias{reprow.matrix}
\alias{reprow.data.frame}
\alias{reprow.data.table}
\title{Data replication function}
\usage{
reprow(x, ...)

\method{reprow}{matrix}(x, ...)

\method{reprow}{data.frame}(x, ...)

\method{reprow}{data.table}(x, ...)
}
\arguments{
\item{x}{matrix, data.frame, data.table.}

\item{...}{times, each}
}
\description{
Data replication function
}
\details{
This function is for replicating data rows
}
\examples{
reprow(iris, times = 3)
reprow(iris, each  = 3)
}
