% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeof.R
\name{sizeof}
\alias{sizeof}
\alias{sizeof.default}
\alias{sizeof.data.frame}
\alias{sizeof.environment}
\title{Object size}
\usage{
sizeof(x, unit)

\method{sizeof}{default}(x, unit = "mb")

\method{sizeof}{data.frame}(x, unit = "mb")

\method{sizeof}{environment}(x, unit = "mb")
}
\arguments{
\item{x}{object vector, data.frame, environment, ...}

\item{unit}{size unit; "B", "KB", "MB", "GB"}
}
\description{
This function calculates the object size
}
\examples{
\dontrun{
sizeof(x = globalenv(), unit = "kb")
}
}
