% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{to_a1_col}
\alias{to_a1_col}
\title{To the a1 style column from an r1c1 style column}
\usage{
to_a1_col(x)
}
\arguments{
\item{x}{a r1c1 style column}
}
\description{
This function coverts an r1c1 style column to the a1 column
}
\examples{
to_a1_col(1)
}
