% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{to_r1c1_col}
\alias{to_r1c1_col}
\title{To the r1c1 style column from an a1 style column}
\usage{
to_r1c1_col(x)
}
\arguments{
\item{x}{a a1 style column}
}
\description{
This function coverts an a1 style column to the r1c1 column
}
\examples{
to_r1c1_col("A")
}
