% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vuw-package.R
\docType{package}
\name{vuw-package}
\alias{vuw}
\alias{vuw-package}
\title{vuw: Virtual Underwriting}
\description{
This package is for a virtual underwriting project.
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
