% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{diff_period}
\alias{diff_period}
\title{Differences between start and end}
\usage{
diff_period(start, end, group = 1)
}
\arguments{
\item{udate}{Date}

\item{sdate}{Date}

\item{ncol}{integer. The number of columns you need to create}

\item{mon}{integer. The months}
}
\value{
Period matrix
}
\description{
Differences between start and end
}
