% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggbar}
\alias{ggbar}
\title{Simple ggplot}
\usage{
ggbar(
  data,
  x,
  y,
  ymin = NULL,
  ymax = NULL,
  ymin_err,
  ymax_err,
  group = NULL,
  color = NULL,
  fill = NULL,
  label,
  family = "Comic Sans MS",
  size = 4,
  angle = 0,
  hjust = 0.5,
  vjust = 0.5
)
}
\description{
Draw a simple bar plot quickly and easily
}
